check

+++ resources/views/admin/elements/_viewaccount.blade.php
@@ -18,13 +18,18 @@
           <td>{{ $item->r_no }}</td>
           <td>{{ $item->branch }}</td>
           <td>
-            <x-modal 
+            @if (Auth::user()->role === 'admin' || Auth::user()->role === 'editor')
+              <x-modal 
               type="editmodal" 
               route="accounts.update,account=>{{ $item->id }}"
-              title="Edit Data" id="{{ $item->id }}" /> ||
+              title="Edit Data" id="{{ $item->id }}" />
+            @endif
+             ||
+            @if (Auth::user()->role === 'admin')
               <button class="btn btn-danger" 
-                onclick="deleteData('{{ route('accounts.destroy',['account'=>$item->id]) }}')">
-              Delete</button>
+              onclick="deleteData('{{ route('accounts.destroy',['account'=>$item->id]) }}')">
+            Delete</button>
+            @endif
           </td>
         </tr>   
       @endforeach
